renamed settings to profile
