--- Cocina/Cocina/Settings/EditPasswordViewController.swift
@@ -1,29 +1,0 @@
-//
-//  EditPasswordViewController.swift
-//  Cocina
-//
-//  Created by Andrea Jimenez on 1/26/24.
-//
-
-import UIKit
-
-class EditPasswordViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .brown
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
